api/requirements.txt
fastapi[standard]>=0.110.0,<0.114.0
uvicorn[standard]>=0.30.0,<0.35.0
google-cloud-aiplatform>=1.43.0
google-cloud-bigquery>=3.18.0
google-cloud-storage>=2.15.0
pandas
yfinance
google-cloud-firestore
